Timing Experiments
Completed code for timing experiments in jupyter notebook.

The first experiment calculates the average amount of time that the simulation
needs to run for multiple number of trials. There are several output
graphs that investigate the amount of time it takes for the code to run.

The second one has to do with the mean blocking and waiting time in the
system. It is desirable to check the number of iterations it takes for
the mean waiting/blocking time to converge.