Coding Markov Chain
Initialised the destination file that will be attempted to find the
exact slution of the problem. Coded so far:
1. Function that creates a networkx object that shows the model to be
built. This is meant to be a preview of the model to ensure that the
correct model is to be built.
2. Function of the state mapping function
3a. Function that creates a symbolic transition matrix (i.e. whose
entries include strings of the symbols (to be updated using sympy)
3b. Function that creates the transition matrix of given parameters