Graph theoretic approach to get pi (#30)
* Minor changes on visualising the Markov model

visualise_ambulance_markov_model() function output plot rotated 90
degrees so that it matches the model created with tikz. networkx class
channged from MultiDiGraph to DiGraph and minor updates on test file as
well.

* Function that finds number of spanning trees

Initial attempt in building a graph theoretic formula to find state
probabilities. A function is investigated that uses the laplacian of a
graph to get the number of spanning trees rooted at a node

* Run black on markov.py