Get python code to generate tikz figure (#31)
* Start building functions for tikz

Two functions are considered where the first one
(convert_networkxx_figure_to_tikz) converts the already existing
networkxx figure into a latex code that generates the corresponding
figure. The second function uses a more logical approach to write down
the latex code for a tikz image where it matches the already existing
approach for building tikz figures.

* Build function that generates latex code for tikz

The function currently generates a string that represents a tikz picture
given the four necesary paramters (num-of-servers, threshold,
system-capacity and parking-capacity). This function is meant to be
utilised in the tex directory to generate markov chains whenever
necessary.

* Create tikz picture using the python code

Addde a tikz picture in the latex code by running the python code and
using the string it returns to perform get the tikz picture.

* Run black on all files

* Add tikzplotlib to requirements

* Minor edits

* Add example tests for the function

* Replace term 1*mu with just mu

* Rerun example in latex

* Update tests

* Run black on tests

* Add docstring for generate_code_for_tikz_figure

* Cherry pick commit 32c559b

Cherry pick a commit from branch investigate-closed-form-of-pi

* Minor edits to test that fails

* Minor typos and edits after supervisory meeting

* Leave out the extra function

Left out the function that turns networkx plot into tikz code. Maybe to
be included at a later stage.