Investigate close form of pi (#29)
* Minor edits after supervisory meeting

* Slightly change sympy symbols of parameters

* Initial attempt of closed-form state probs formula

Used sympy to investigate a potential relation among the state
probabilities of the markov chain. Managed to get expressions of smaller
models and try to generalise these expressions to fit larger models. An
intial attempt of looking at the ratios of state probabilities of
adjacent states was made.

* Investigate the probability of being in (0,0)

Looked into small models and tried to get an expression of the numerator
of the probability of being at state (0,0) and then attempted to
generalaise that expression for larger models. The expression is not yet
completed but a lot of progress has been made.

* Run black on python file of examples

* Graph theoretic approach to get pi (#30)

* Minor changes on visualising the Markov model

visualise_ambulance_markov_model() function output plot rotated 90
degrees so that it matches the model created with tikz. networkx class
channged from MultiDiGraph to DiGraph and minor updates on test file as
well.

* Function that finds number of spanning trees

Initial attempt in building a graph theoretic formula to find state
probabilities. A function is investigated that uses the laplacian of a
graph to get the number of spanning trees rooted at a node

* Run black on markov.py

* Apply Markov matrix tree theorem and rearrange some cells

* Add example model with a system capacity of 9

* Run black

* Store outcome in jupyter notebook

* Update Black

* Function that genetrates P00 combinatorically