Fix proportion formula to use time in the system
Two approximations were used on the Erlang distributions and then used
the cdf of the Hypoexponential distribution. The default way to be used
will be the Hyopexponential distribution.

All three comparisons plots have been created (class 1, class 2, and
overall). In addition these comparisons were ran for different values
of parameters.