Fix proportions formula for when C=1
When C=1 (i.e. there is only one server) the Hypoexponential
distribution should not be used. In such a case the Erlang distribution
should be used because the parameters of the exponential distributions
that are summed together are the same.

In essence, Hypoexp. should be used for (λ_1, λ_2, ..., λ_n) when there
is at leats one distinct pair of λ_i's. Otherwise, if λ_i = λ, for all
i, the Erlang distribution should be used instead. And that happens only
when C=1.
=> (Cμ, Cμ, ..., μ) -> (μ, μ, ..., μ)