Get specific function of psi
The general psi function is used in the cdf of the Hypoexponential
distribution and it requires computing a derivative of a function.
After investigating the outcome of that derivative, an alternative
closed-form expression has been found that works only for the
specific Markov problem.

More specifically the new function calculates the cdf of the
Hypoexponential distribution when there only two distinct parameters.
In essence, it gets the cdf of Hypo(λ1, λ2, ..., λn) where:
{λ1, ..., λ(n-1) = Cμ} and {λn = μ}