Waiting time probability distribution (#42)
* Formula of percentage of individuals within target

Jupyter notebook that contains the formula for the percentage of
individuals within a specified target. The notebook contains the
formula for each of the two classes and the overall one.

* Fix proportion formula to use time in the system

Two approximations were used on the Erlang distributions and then used
the cdf of the Hypoexponential distribution. The default way to be used
will be the Hyopexponential distribution.

All three comparisons plots have been created (class 1, class 2, and
overall). In addition these comparisons were ran for different values
of parameters.

* Add parameter min_threshold

* Fix proportions formula for when C=1

When C=1 (i.e. there is only one server) the Hypoexponential
distribution should not be used. In such a case the Erlang distribution
should be used because the parameters of the exponential distributions
that are summed together are the same.

In essence, Hypoexp. should be used for (λ_1, λ_2, ..., λ_n) when there
is at leats one distinct pair of λ_i's. Otherwise, if λ_i = λ, for all
i, the Erlang distribution should be used instead. And that happens only
when C=1.
=> (Cμ, Cμ, ..., μ) -> (μ, μ, ..., μ)

* Get specific function of psi

The general psi function is used in the cdf of the Hypoexponential
distribution and it requires computing a derivative of a function.
After investigating the outcome of that derivative, an alternative
closed-form expression has been found that works only for the
specific Markov problem.

More specifically the new function calculates the cdf of the
Hypoexponential distribution when there only two distinct parameters.
In essence, it gets the cdf of Hypo(λ1, λ2, ..., λn) where:
{λ1, ..., λ(n-1) = Cμ} and {λn = μ}

* Minor adjustment in Markov visualisation function

* Add tests and minor edits for comparisons.py file

* Add additional tests for waiting.py functions

Some of the functions in waiting.py didn't have tests. This commit
adds some tests in test_markov_waiting.py that tests the functions
that didn't have tests

* Use ciw version 2.1.0 in github actions

* Add additional tests for utils.py functions

* Edit tests to improve coverage results

* Run isort to sort imported libraries

* Remove exponential distribution cdf

Exponential distribution Exp(λ) changed to Erlang distribution
Erlang(1, λ).

* Attempt to proof closed-form of hypo. formula

Initial attempt on trying to prove the specific closed-form (for k=1 and
k=2) formula. Tried to get it by proving it by induction. Next attempt
will include using the Taylor Series expansion.

* Add tests for new utils.py functions

* Modularise utils.py functions

Modularise the following two functions:
- get_probability_of_accepting
- get_proportion_of_individuals_not_lost

* Update tests and fix utils.py module

* Add tests for new simulation.py functions

* Modularise simulation.py functions

* Update simulation tests

* Run black

* Proof of specific hypoexponential formula

- Update exiting specific hypoexponential formula in jupyter notebook
- Jupyter notebook containing the proof of the specific
hypoexponential formula

* Add tests for proportion.py module

* Modularise with docstrings proportion.py functions

* Minor changes in jupyter notebook

* Update proportion functions tests

* Add failing tests for comparisons.py functions

* Modularise comparison functions of formula

* Remove unused imports

* Restructure comparisons.py module

- Combine functions that output a comparison plot into one
- Rename function to make sense for proportion formula
- Rename a times_to_compare to measure_to_compare

* Run black

* Fix function that did not work on python 3.6

np.math.prod() was introduced in python3.8 and is replaced so
that the code works on python3.6

* Add another and modify existing example model

* Write sections on erlang and hypoexponential

Start writing section on getting the mean proportion of arrivals within
time target. Four subsections are written so far:
- Distribution of time at specific state (1 server case)
- Distribution of time at specific state (with multiple servers)
- Specific cdf of hypoexponential distribution
- Proof of equation specific cdf of hypo.

* Write proportion within target sections

Write section for class 1 and class 2 individuals and another section
for the overall proportion of both classes using lost individuals.

* Update .gitignore file

* Update bibliography.bib file and add citations

* Minor changes to paragraph

* Apply suggestions from supervisory meeting

* Minor edits