Add script for generating data (#46)
* Add README.md for script

* Fix minor bug in blocking.py

Fix an issue that caused the code to show an error when trying to get
the blocking time for a system with a buffer capacity of 1 and a system
capacity equal to the threshold.

* Write script for generating data

First draft of the script. Still need to change the way we iterate over
different parameters, remove the cache from  get_payoff_matrices( )
and maybe change how arrays are saved in the .csv file

* Remove lru_cache from get_payoff_matrices function

The lru_cache( ) is temporarily removed because there seems to have a
conflict when giving a numpy.array as an argument.

* Use np.array2string to sav arrays to csv file

* Changes the way to iterate over some parameters

Every random system that is generated will now iterate over different
values of lambda_2, lambda_1_1, lambda_1_2 and alpha before changing
the other parameters.

* Change how we iterate over parameters

Changes made:
- Max lambda_2 value is doubled
- Increased linspace num parmater
- Add target as one of the main parameters to iterate over
- Some other minor edits

* Cerate brand new script to generate data for game

* Add documentation on all functions

* Round parmateres to 2 decimals

Choosing to round all parametres to 2 decimals so that the direcotries
created are more readable. I decided to round the true values as well so
that they are the same with the directories name.

* Remove column/row names on generated .csv files

* Parallelise get_payoff_matrices function

- Restructured the way we calculated the routing matrix and the payoff
matrices so that the whole procedure can be parallelised.
- Fixed the tests as well.

* Add dask to requirements.txt and environment.yml

* Update requirements.txt

Had to specify dask[complete] so that the delayed attribute
is downloaded as well.

* Update main.py to use multiple processes with dask

* Change the way the data are saved - numpy.savetxt

Used to convert the array into a panda object and then use
DataFrame.to_csv to save the file in a .csv file. Instead numpy's
savetxt function is used now.

* Pass num of processes as a command line argument

Use sys.argv to get the arguments passed in the command line and
if there is only one argument passed that can be converted to an
integer then run main.py with that many dask workers.

* Update environment.yml

* Minor refactoring in src/ambulance_game/game.py

Co-authored-by: Vince Knight <vincent.knight@gmail.com>

* Run black

* Update the structure for the data subdirecotries

Use numpy's savez_compressed() function to store all
three arrays in one compressed file main.npz

* Fix cache in main.py

The cache that was not working as expected. Fixed it by changing
the dictionary entry into a tuple.

* Minor fixes

Remove unnecessary function and edit a docstring

* Change README.md files

Update the readme file that is located in data/game_data/ and
add instructions of how to read the generated data to the readme
files that are created by main.py.

* Add the generated data in .gitignore

* Add addtional tests for the new functions

Add some tests for the functions that were created to make the
source code parallelised

* Fix minor bug in test

Use np.allclose to compare values of tuples

* Round some parameter values to one decimal

Some parameters values have been rounded to 1 decimal place but this
may have to change back.

* Fix readme contents

Avoid using the word simply in instructions

* Use hypothesis in test_game.py

Remove random library from tests and use hypothesis insetad.

* Minor changes in script

Stop rounding floats and change parameters/ directory into
_parameters/ so that it appears first

* Use a hash for the subdirectory names

Used hashlib to create a hash for the given set of parameters and payoff
matrices. Use the hash as the subdirecotry names and include it in the
parameters/main.csv file as well. Update docstrings and README.md files
accordingly.

* Add function that creates hash

* Update README.md contents

* Change data generation parameters

Point the script at a different set of parameters to explore, and
decrease the number of iterations for each parameter so that fewer
cases are explored.

Co-authored-by: Vince Knight <vincent.knight@gmail.com>