Include setup.py again
The main (and only) reason to add setup,py again in the repository is
to be able to make editable imports of the library.
While pyproject.toml is the new default now, setup.py is used when
making changes locally so that it's not needed to run `pip install .`
every time.

What is not great about this is that there are now 3 duplicate sets
of library versions (setup.py, pyproject.toml, environment.yml).