Add script for data analysis (#49)
* Create .py file to extract performance measures

The functions created will be used on the script that will be run on
siren to analyse the generated data.

* Create .ipynb files with preliminary analyses

Jupyter notebooks for some of the experiments. Performance values like
the price of anarchy and nash equilibrias are investigated.

* Change the way price of anarchy is calculated

An array is created that consists all performance measures and is
multiplied by the equilibrium strategies to find the performance
measure for that set of strategies.

* Update .gitignore file

* Refactor python file and improve readability

* Update nashpy in environment.yml

Include nashpy version 0.0.23 which is the latest one that
includes asymmetric replicator dynamics as well

* Rename jupyter notebook

* Jupyter notebook of price of anarchy analysis

Jupyter notebook containing some analysis of the price of anarchy of
blocking times, waiting times, and lost individuals.

* Add jupyter nb that explores PoA inefficiencies

The Jupyter notebook explores how the price of anarchy can actually
change when we change the actual game. This can be achieved by either
increasing the arrival rate, increasing the system capacity or incentivise
the players by changing the utilities of the matrix.

* Add additional notebooks of PoA analysis

* Add scripts to generate data in their own dir

* Minor changes in poa analysis functions

* Add new function in ambulance_game

New function: get_accepeting_proportion_of_individuals

* Update packaging (#51)

* Add CHANGELOG.md

* Install and run flit

pyproject.toml generated

* Start using flake8 and configure it in .flake8

* Apply changes based on flake8 suggestions

* Configure mypy in pyproject.toml

* Apply pylint's suggestions to source code

* Apply pylint's suggestions to tests

* Add module docstrings to all source and test files

* Update hypothesis range on a test

* Include libraries in pyproject.toml

In addition upgraded scipy from version 1.4.1 to version 1.7.1
due to a clash with python 3.9

* Remove python 3.6, 3.7 from CI and add python 3.9

* Add max_examples argument in a test

* Update dask and matplotlib libraries

* Include tox.ini file

This file will contain all commands that will be executed in
github actions

* Modify config.yml to run tox

* Remove setup.py

* Change version of numpy in pyproject.toml

* Update environment.yml

* Disable failing tests for MacOS - py39

* Remove requirements.txt

* Apply flake8 and pylint on tests

* Add dist  directory to .gitignore

* Remove unnecessary line in tox.ini

* Include setup.py again

The main (and only) reason to add setup,py again in the repository is
to be able to make editable imports of the library.
While pyproject.toml is the new default now, setup.py is used when
making changes locally so that it's not needed to run `pip install .`
every time.

What is not great about this is that there are now 3 duplicate sets
of library versions (setup.py, pyproject.toml, environment.yml).

* Update README.md file

* Make sure that test coverage is 100% (#52)

* Include not tested lines in coverage report

* Add test for comparisons.py

* Add docstring to test

* Add test for expected error

* Add missing tests for markov.py

* Add tests for proportion.py

* Add tests for utils.py

* Add/fix tests for simulation.py

* Minor change to test name

* Force checks to fail when coverage is below 100%

* Specify dtype=int to remove warnings

* Stop skipping one test

* Update tests to use np.linalg.solve

numpy.linalg.lstsq might have an issue when running on python3.9 on Mac

* Exclude Ubuntu and Windows from CI (test run)

* Fix minor issue

Skip test of np.linalg.lstsq on Mac + py39 and exclude the fuction
in the source code from coverage.

* Remove pytest.mark.skipif from test

Re-enable a skipped test that I missed in the previous commit

* Include again Windows and Ubuntu to CI

* Modify packaging (#53)

* Remove setup.py

* Update README.md to include flit usage

Add instructions on how to install flit and how to use it to install
an editable install of the library.

* Fix badge for issues