Make game compatible with simulation (#63)
* Remove duplicate function

* Enable result collection from simulation objects

Two new functions created:
1. Analyses simulation objects to get results
2. Analyses simulation objects and gets state probabilities

Both (1) and (2) were done in the past by resimulating the model. This
is not necessary now. In addition the Queue class and the tests are
updated accordingly.

* Run black on simulation.py

* Add lost inds to decision values

The function get_mean_blocking_difference_between_two_systems now uses
the weighted average of the blocking time and the proportion of lost
individuals rather than just the blocking time (similar to markov).

* Change function names

Make function names consistent across simulation and markov methods

* Refactor source code and tests

The source code and tests have been changed in such a way
so that the game.py module can run with the simulation instead
of just the markov chain model.

* Add one extra test for simulation

* Remove special character in docstring

* Allow no arrivals in simulation

Add functionality such that lambda_1 and/or lambda_2
can be 0.

* Minor change to test

This test sometimes failed when hypothesis produced an error of:
Unreliable test timings. Fix this by setting deadline to None.