Update state dependent distribution logic
Fixed a faulty logic of the state_(server)_dependent distribution of the
simulation. When an exponentially distributed random number is
sampled, it is sampled based on the current state of the model.
Because of the way `ciw` is written, in my case that wasn't always the
case. This change fixes that inconsistency. Basically instead of using
ciw's statetracker to get the current state we do that manually by
looking in the simulation.