Reinforcement learning experiments (attempt 3) (#68)
* Copy previous scripts to be changed

* Update state dependent distribution logic

Fixed a faulty logic of the state_(server)_dependent distribution of the
simulation. When an exponentially distributed random number is
sampled, it is sampled based on the current state of the model.
Because of the way `ciw` is written, in my case that wasn't always the
case. This change fixes that inconsistency. Basically instead of using
ciw's statetracker to get the current state we do that manually by
looking in the simulation.

* Update tests

* Calculate state probabilities on every iteration

Several things have been addresed in this commit:
- Added state probabilities calculation on every iteration
- Increased the upper bound for a server's rate
- Can now give additional inputs through the command line
- File directory is now calculated with the values of the parameters

* Notebook used to investigate issue with RL

* Initialise notebook to be used to analyse results

* Add python plotting scripts

* Add jupyter notebooks with plots of results

* Add experiments for INFORMS/CORS talk

* Apply suggestions from pylint

* Add experiments with unbounded rates

* Apply pylint suggestions to tests

* Include some additional plots of results

Two nice plots added:
- Plot different strating mu in same graph
- Add 3rd plot of weighted mean of ratio of service rate to max rate