Implement RL algorithm for servers behaviour (#64)
* Add option to not use the lru cache

* Update ciw version

In the new version of ciw, server priorities are implemented. I can now
input a custom server priority function to implement fair allocation of
individuals among servers. Also move get_arrival_distribution function
fom simulation.py to dists.py.

* Add upper limit to pyproject requirements

* Set nashpy to version 0.0.23

* Add test for fair allocation of individuals

* Minor refactoring of server priorities

Allow for custom functions to be input as server priority functions

* Update black library and reformat files

* Modify some notebooks for supervisory meeting

* Add notebook on server utility functions

This jupyter notebook contains some possible utility functions
that may be used for the reinforcement learning algorithm

* Parametrise using parameters from paper

Use parameters and observed behaviour of servers that is described in
the paper "Behaviours in Healthcare Queueing Systems" to observe the
values of the utility functions

* Initial implementation of utility functions

Currently this does not support multiple trials of the simulation. The
next commit will aim to address that

* Set the deadline to "None" for a particular test

This test took some time to run and because of that caused
Hypothesis to produce an error.

* Add graph from paper to jupyter notebook

* Use multiple trials in utility functions

Utility functions now get a list of simulation objects and calculate
the utilities based on the average performance measures of all
simulations. Old implementation only took one simulation object

* Update server attributes as simulation is running

Add 2 new attributes to every Server object while the simulation is
running. First attribute is "served_inds" which contains a list of all
times that any individual has been served by that server.
Second attribute is "service_times" which is a list of all service times
for that server.

* Initial run of reinforcement learning and results

Two jupyter notebooks with some runs of the allgorithm and a couple of
.csv files containing the results from these runs. The produced graphs
on the notebooks can be replicated using the .csv files

* Run black on test file

* Reinforcement learning experiments (#65)

* Write reinforcement learning scripts

Four scripts are created for the following values of e: 0, 0.1, 0.5, 1

* Add jupyter notebook of results

Jupyter notebook that consists of several plots to visualise the
reinforcement learning algorithm's results. i.e. plots of the utilities
and rates of each server

* Write more scripts for the RL algorithm

* Update .gitignore file

* Add jupyter notebook of further experiments

* Reinforcement learning experiments (attempt 2) (#66)

* New utility function

Weighted average between mean idle proportion time and mean accpted
individuals.

* Restructure RL data directory

* Add README for reinforcement learning experiments

* Add script for new experiments

New experiments use utility function 7

* Remove commented out parameters

* Add initial version of base jupyter notebook

* Add plots for the new data

* Improve transition matrix implementation (#67)

* Improve the transition matrix implementation

Rather than iterating through all possible pairs of states to get the
transition matrix now we just find all non-zero positions and just
fill those. This change can be very useful when we try to get
performance measures for queues with larger capacities.

* Apply minor suggestion based on black

* Add test for the old transition matrix function

* Add function to reconstruct rates

Two functions are added; one to reconstruct the dictionary of rates from
raw rates file and one to reconstruct the matrix of rates from the
dictionary

* Add a jupyter notebook of further analysis

* Reinforcement learning experiments (attempt 3) (#68)

* Copy previous scripts to be changed

* Update state dependent distribution logic

Fixed a faulty logic of the state_(server)_dependent distribution of the
simulation. When an exponentially distributed random number is
sampled, it is sampled based on the current state of the model.
Because of the way `ciw` is written, in my case that wasn't always the
case. This change fixes that inconsistency. Basically instead of using
ciw's statetracker to get the current state we do that manually by
looking in the simulation.

* Update tests

* Calculate state probabilities on every iteration

Several things have been addresed in this commit:
- Added state probabilities calculation on every iteration
- Increased the upper bound for a server's rate
- Can now give additional inputs through the command line
- File directory is now calculated with the values of the parameters

* Notebook used to investigate issue with RL

* Initialise notebook to be used to analyse results

* Add python plotting scripts

* Add jupyter notebooks with plots of results

* Add experiments for INFORMS/CORS talk

* Apply suggestions from pylint

* Add experiments with unbounded rates

* Apply pylint suggestions to tests

* Include some additional plots of results

Two nice plots added:
- Plot different strating mu in same graph
- Add 3rd plot of weighted mean of ratio of service rate to max rate